requirements.txt
torch==1.6.0
torchvision==0.7.0
numpy==1.18.5
tqdm==4.47.0
Pillow==7.2.0
wandb==0.10.11
gensim==3.8.3